+++ plugin/src/main/java/net/dzikoysk/funnyguilds/feature/hooks/placeholderapi/PlaceholderAPIHook.java
@@ -7,7 +7,7 @@ import net.dzikoysk.funnyguilds.FunnyGuilds;
 import net.dzikoysk.funnyguilds.config.PluginConfiguration;
 import net.dzikoysk.funnyguilds.feature.hooks.AbstractPluginHook;
 import net.dzikoysk.funnyguilds.feature.placeholders.TablistPlaceholders;
-import net.dzikoysk.funnyguilds.feature.placeholders.impl.Placeholder;
+import net.dzikoysk.funnyguilds.feature.placeholders.placeholder.Placeholder;
 import net.dzikoysk.funnyguilds.feature.prefix.IndividualPrefix;
 import net.dzikoysk.funnyguilds.guild.Guild;
 import net.dzikoysk.funnyguilds.rank.RankUtils;
