+++ plugin/src/main/java/net/dzikoysk/funnyguilds/feature/placeholders/GuildPlaceholders.java
@@ -10,8 +10,8 @@ import net.dzikoysk.funnyguilds.FunnyGuilds;
 import net.dzikoysk.funnyguilds.config.MessageConfiguration;
 import net.dzikoysk.funnyguilds.config.NumberRange;
 import net.dzikoysk.funnyguilds.config.PluginConfiguration;
-import net.dzikoysk.funnyguilds.feature.placeholders.impl.GuildPlaceholder;
-import net.dzikoysk.funnyguilds.feature.placeholders.impl.Placeholder;
+import net.dzikoysk.funnyguilds.feature.placeholders.placeholder.GuildPlaceholder;
+import net.dzikoysk.funnyguilds.feature.placeholders.placeholder.Placeholder;
 import net.dzikoysk.funnyguilds.feature.placeholders.resolver.MonoResolver;
 import net.dzikoysk.funnyguilds.feature.placeholders.resolver.PairResolver;
 import net.dzikoysk.funnyguilds.guild.Guild;
