+++ plugin/src/main/java/net/dzikoysk/funnyguilds/feature/placeholders/placeholder/Placeholder.java
@@ -1,4 +1,4 @@
-package net.dzikoysk.funnyguilds.feature.placeholders.impl;
+package net.dzikoysk.funnyguilds.feature.placeholders.placeholder;
 
 import java.util.Objects;
